+++ models/fence.go
@@ -9,7 +9,7 @@ import (
 
 var (
 	UpgradeTypesRadius = [...]int{100, 150, 200, 250, 300, 350, 400}
-	UpgradeTypesRent   = [...]float32{1, 1.5, 2, 2.5, 3, 3.5, 4}
+	UpgradeTypesRent   = [...]float64{1, 1.5, 2, 2.5, 3, 3.5, 4}
 	FenceMaxTTL        = 60 * 60 * 24 * 7
 	FenceMinRadius     = maths.Min(UpgradeTypesRadius[:])
 	FenceMaxRadius     = maths.Max(UpgradeTypesRadius[:])
@@ -23,7 +23,7 @@ type Fence struct {
 	Lat            float64   `json:"centerLat"`
 	Lon            float64   `json:"centerLon"`
 	Radius         int       `json:"radius"`
-	RentMultiplier float32   `json:"rentMultiplier"`
+	RentMultiplier float64   `json:"rentMultiplier"`
 	TTL            int       `json:"ttl"`
 	DiesAt         time.Time `json:"diesAt"`
 	Name           string    `json:"name"`
